推Parser.cs

+++ shared/src/IF/I_Segment.cs
@@ -4,11 +4,11 @@ public interface I_Segment<T>{
 	/// <summary>
 	/// 含
 	/// </summary>
-	public u8 Start{get;set;}
+	public u64 Start{get;set;}
 	/// <summary>
 	/// 含
 	/// </summary>
-	public u8 End{get;set;}
+	public u64 End{get;set;}
 
 	public T Data{get;set;}
 }
